[WIP] All Tests passing in mock backend

+++ ServerlessBackendAWS/services/empresa-service/src/sidecar.ts
@@ -1,16 +1,21 @@
 import { Context, APIGatewayProxyEvent, Callback, APIGatewayProxyResult } from 'aws-lambda';
 import { BusinessEvent, InfrastructureMetric, ProxyResolver, ServiceError } from './types.d';
-import { AwsHelper } from  './aws-helper';
+import { AwsHelper } from './aws-helper';
 import { DynamoDB } from 'aws-sdk';
+import { writeFileSync, existsSync, mkdirSync } from 'fs';
+import { ResponseBuilder } from './utilities';
+
+const isLambda = !!(process.env.LAMBDA_TASK_ROOT || false);
 
 enum SidecarError {
     UserNotAuthorized = 'UserNotAuthorized',
+    ResourceNotFound = 'ResourceNotFound',
     InvalidObjectBody = 'InvalidObjectBody',
     InvalidOperation = 'InvalidOperation',
-    ResourceNotAvailable = 'ResourceNotAvailable',
+    InvalidConfiguration = 'InvalidConfiguration',
+    DependencyNotAvailable = 'DependencyNotAvailable',
     DependencyError = 'DependencyError',
     ConfigurationNotAvailable = 'ConfigurationNotAvailable',
-    InvalidConfiguration = 'InvalidConfiguration',
     undefined = 'undefined'
 }
 
@@ -29,7 +34,7 @@ export class Sidecar {
      * be verified by AWS Security Services.
      */
     userIsAuthorizedToPerformOperation(event: APIGatewayProxyEvent): Promise<boolean> {
-        
+
         const trace = new OperationBuilder().withAction('userIsAuthorizedToPerformOperation');
 
         return new Promise((resolve: Function, reject: Function) => {
@@ -97,23 +102,31 @@ export class Sidecar {
      */
     persistOnLake(bucketName: string, eventId: string, event: APIGatewayProxyEvent): Promise<boolean> {
 
-        const parameters = {bucketName: bucketName, eventId: eventId};
+        const parameters = { bucketName: bucketName, eventId: eventId };
         const trace = new OperationBuilder().withAction('persistOnLake').withPayload(event.body).withParameters(parameters);
 
         const bucket = this.dependencyResolver.bucket('region', bucketName);
         return new Promise((resolve: Function, reject: Function) => {
 
-            bucket.putObject(eventId, JSON.stringify(event)).then(success => {
-    
-                this.trail.push(trace.build());
-                resolve(true);            
-    
-            }).catch(error => {
+            if (bucket) {
 
-                this.trail.push(trace.withError(error).build());
+                bucket.putObject(eventId, JSON.stringify(event)).then(success => {
+                    this.trail.push(trace.build());
+                    resolve(true);
+                }).catch(error => {
+                    this.trail.push(trace.withError(error).build());
+                    reject(error);
+                });
+
+            } else {
+
+                const parameters = { resourceType: 'bucket', resourceName: bucketName }
+                const error: ServiceError = {code: SidecarError.DependencyNotAvailable, httpStatusCode: 500, resource: bucketName, payload: event};
+                this.publishMetric({ metricArn: error.code, value: 1 })
+                this.trail.push(trace.withError(error).withParameters(parameters).build());
                 reject(error);
 
-            });
+            }
 
         });
 
@@ -128,24 +141,44 @@ export class Sidecar {
      */
     getInTable(tableName: string, key: string): Promise<Object> {
 
-        const parameters = {tableName: tableName, key: key};
+        const parameters = { tableName: tableName, key: key };
         const trace = new OperationBuilder().withAction('getInTable').withParameters(parameters);
 
         const table = this.dependencyResolver.table('region', tableName);
         return new Promise((resolve: Function, reject: Function) => {
 
-            table.getItem(key).then((object: DynamoDB.AttributeMap) => {
-                const result = AwsHelper.unmarshalObject(object);
-                this.trail.push(trace.build());
-                resolve(result);
-            }).catch((error: ServiceError) => {
-                const parameters = { itemId: key }
-                this.trail.push(trace.withParameters(parameters).withError(error).build());
+            if (table) {
+
+                table.getItem(key).then((object: DynamoDB.AttributeMap) => {
+
+                    if (object) {
+                        const result = AwsHelper.unmarshalObject(object);
+                        this.trail.push(trace.build());
+                        resolve(result);
+                    } else {
+                        const error: ServiceError = {code: SidecarError.ResourceNotFound, httpStatusCode: 404, resource: tableName + '/' + key};
+                        this.trail.push(trace.withError(error).withParameters(parameters).build());
+                        reject(error);
+                    }
+
+                }).catch((error: ServiceError) => {
+                    const parameters = { itemId: key }
+                    this.trail.push(trace.withParameters(parameters).withError(error).build());
+                    reject(error);
+                });
+
+            } else {
+
+                const parameters = { resourceType: 'table', resourceName: tableName }
+                const error: ServiceError = {code: SidecarError.DependencyNotAvailable, httpStatusCode: 500, resource: tableName, payload: event};
+                this.publishMetric({ metricArn: error.code, value: 1 })
+                this.trail.push(trace.withError(error).withParameters(parameters).build());
                 reject(error);
-            });
+
+            }
 
         });
-    
+
     }
 
     /**
@@ -160,44 +193,56 @@ export class Sidecar {
     persistInTable(tableName: string, key: string, event: APIGatewayProxyEvent): Promise<boolean> {
 
         const operation = event.httpMethod;
-        const parameters = {tableName: tableName, key: key, operation: operation};
+        const parameters = { tableName: tableName, key: key, operation: operation };
         const trace = new OperationBuilder().withAction('persistInTable').withPayload(event.body).withParameters(parameters);
 
         const table = this.dependencyResolver.table('region', tableName);
         return new Promise((resolve: Function, reject: Function) => {
 
-            if (operation.toLowerCase() === 'post' || operation.toLowerCase() === 'put') {
-                
-                const payload = AwsHelper.marshalObject(event.body);
-                table.putItem(key, payload).then(success => {
-                    this.trail.push(trace.build());
-                    resolve(true);            
-                }).catch(error => {
-                    this.trail.push(trace.withError(error).build());
-                    reject(error);
-                });
+            if (table) {
 
-            } else if (operation.toLowerCase() === 'delete') {
-                
-                table.deleteItem(key).then(success => {
-                    this.trail.push(trace.build());
-                    resolve(true);            
-                }).catch(error => {
-                    this.trail.push(trace.withError(error).build());
+                if (operation.toLowerCase() === 'post' || operation.toLowerCase() === 'put') {
+
+                    const payload = AwsHelper.marshalObject(event.body);
+                    table.putItem(key, payload).then(success => {
+                        this.trail.push(trace.build());
+                        resolve(true);
+                    }).catch(error => {
+                        this.trail.push(trace.withError(error).build());
+                        reject(error);
+                    });
+
+                } else if (operation.toLowerCase() === 'delete') {
+
+                    table.deleteItem(key).then(success => {
+                        this.trail.push(trace.build());
+                        resolve(true);
+                    }).catch(error => {
+                        this.trail.push(trace.withError(error).build());
+                        reject(error);
+                    });
+
+                } else {
+
+                    const parameters = { requestedOperation: operation }
+                    const error: ServiceError = {code: SidecarError.InvalidOperation, httpStatusCode: 405, resource: tableName, payload: event};
+                    this.trail.push(trace.withParameters(parameters).withError(error).build());
                     reject(error);
-                });
+
+                }
 
             } else {
 
-                const parameters = { requestedOperation: operation }
-                const error = SidecarError.InvalidOperation;
-                this.trail.push(trace.withParameters(parameters).withError(error).build());
+                const parameters = { resourceType: 'table', resourceName: tableName }
+                const error: ServiceError = {code: SidecarError.DependencyNotAvailable, httpStatusCode: 500, resource: tableName, payload: event};
+                this.publishMetric({ metricArn: error.code, value: 1 })
+                this.trail.push(trace.withError(error).withParameters(parameters).build());
                 reject(error);
 
             }
 
         });
-    
+
     }
 
     /**
@@ -214,18 +259,30 @@ export class Sidecar {
 
         return new Promise((resolve: Function, reject: Function) => {
 
-            topic.publish(businessEvent.body).then(messageId => {
+            if (topic) {
 
-                const parameters = { messageId: messageId }
-                this.trail.push(trace.withParameters(parameters).build());
-                resolve(true);
+                topic.publish(businessEvent.body).then(messageId => {
 
-            }).catch(error => {
+                    const parameters = { messageId: messageId }
+                    this.trail.push(trace.withParameters(parameters).build());
+                    resolve(true);
 
-                this.trail.push(trace.withError(error).build());
+                }).catch(error => {
+
+                    this.trail.push(trace.withError(error).build());
+                    reject(error);
+
+                });
+
+            } else {
+
+                const parameters = { resourceType: 'topic', resourceName: topicArn }
+                const error: ServiceError = {code: SidecarError.DependencyNotAvailable, httpStatusCode: 500, resource: topicArn, payload: event};
+                this.publishMetric({ metricArn: error.code, value: 1 })
+                this.trail.push(trace.withError(error).withParameters(parameters).build());
                 reject(error);
 
-            });
+            }
 
         });
 
@@ -237,53 +294,40 @@ export class Sidecar {
      * @description
      * Publish Metric Value
      */
-    publishMetric(metricValue: InfrastructureMetric): Promise<boolean> { 
+    publishMetric(metricValue: InfrastructureMetric): Promise<boolean> {
 
         const trace = new OperationBuilder().withAction('publishMetric').withPayload(JSON.stringify(metricValue))
         const metric = this.dependencyResolver.metric('region', metricValue.metricArn);
         return new Promise((resolve: Function, reject: Function) => {
 
-            metric.publish(metricValue.metricArn, JSON.stringify(metricValue.value))
-            .then(success => {
+            if (metric) {
 
-                this.trail.push(trace.build());
-                resolve(true);
+                metric.publish(metricValue.metricArn, JSON.stringify(metricValue.value))
+                    .then(success => {
 
-            }).catch(error => {
+                        this.trail.push(trace.build());
+                        resolve(true);
 
-                this.trail.push(trace.withError(error).build());
+                    }).catch(error => {
+
+                        this.trail.push(trace.withError(error).build());
+                        reject(error);
+
+                    });
+
+            } else {
+
+                const parameters = { resourceType: 'metric', resourceName: metricValue.metricArn }
+                const error: ServiceError = {code: SidecarError.DependencyNotAvailable, httpStatusCode: 500, resource: metricValue.metricArn, payload: event};
+                this.publishErrorInContingence(JSON.stringify(error))
+                this.publishErrorInContingence(JSON.stringify({ metricArn: error.code, value: 1 }))
+                this.trail.push(trace.withError(error).withParameters(parameters).build());
                 reject(error);
 
-            });
+            }
 
         });
 
-    }
-
-    /**
-     * @param event APIGatewayProxyEvent
-     * @returns APIGatewayProxyResult
-     * @description
-     * Create Response Object based on event received
-     */
-    createResponse(event: APIGatewayProxyEvent, operation: string, content: Object): APIGatewayProxyResult {
-        
-        const result = {statusCode: 200, body: JSON.stringify(event)}
-        return result;
-    
-    }
-
-    /**
-     * @param error SidecarError
-     * @returns APIGatewayProxyResult
-     * @description
-     * Create Response Object based on throwed error
-     */
-    createErrorResponse(error: ServiceError): APIGatewayProxyResult {
-        
-        const result = {statusCode: error.httpStatusCode, body: ''}
-        return result;
-    
     }
 
     /**
@@ -293,12 +337,67 @@ export class Sidecar {
      * Publish Error in Configured Metric/LogGroup/LogStream/LogConcentrator
      */
     publishError(error: ServiceError, payload: string): Promise<ServiceError> {
-        
+
+        console.log('publishing error')
+
+        let e = new Error();
+        let frame = e.stack.split("\n")[2];
+        let lineNumber = frame.split(":")[1];
+        let functionName = frame.split(" ")[5];
+
+        const parameters = { function: functionName, line: lineNumber, error: error }
+        const trace = new OperationBuilder().withAction('publishError').withParameters(parameters).withPayload(JSON.stringify(payload))
         return new Promise((resolve: Function, reject: Function) => {
 
+            this.trail.push(trace.build());
             resolve(error);
 
         });
+    }
+
+    /**
+     * @param event APIGatewayProxyEvent
+     * @returns APIGatewayProxyResult
+     * @description
+     * Create Response Object based on event received
+     */
+    createResponse(event: APIGatewayProxyEvent, content: Object, objectId: string): APIGatewayProxyResult {
+
+        var result:APIGatewayProxyResult
+        const operation = event.httpMethod.toLowerCase()
+
+        if (operation == 'post') { 
+
+            const url = event.path + '/' + objectId;
+            result = ResponseBuilder.created(url, content)
+
+        } else if (operation == 'delete') {result = ResponseBuilder.ok(null);
+        } else if (operation == 'put' || operation == 'get') {result = ResponseBuilder.ok(content);}
+
+        return result;
+
+    }
+
+    /**
+     * @param error SidecarError
+     * @returns APIGatewayProxyResult
+     * @description
+     * Create Response Object based on throwed error
+     */
+    createErrorResponse(error: ServiceError): APIGatewayProxyResult {
+
+        switch (error.code) {
+            case 'UserNotAuthorized': return ResponseBuilder.forbidden(error.resource, ''); 
+            case 'ResourceNotFound': return ResponseBuilder.notFound(error.resource, ''); 
+            case 'InvalidObjectBody': return ResponseBuilder.badRequest(error.resource, '');
+            case 'InvalidOperation' : return ResponseBuilder.badRequest(error.resource, '');
+            case 'InvalidConfiguration': return ResponseBuilder.internalError(error.resource, '');
+            case 'DependencyNotAvailable': return ResponseBuilder.internalError(error.resource, '');
+            case 'DependencyError': return ResponseBuilder.internalError(error.resource, '');
+            case 'ConfigurationNotAvailable': return ResponseBuilder.internalError(error.resource, '');
+            default: return ResponseBuilder.internalError(error.resource, '');
+        }
+
     }
 
     /**
@@ -308,10 +407,10 @@ export class Sidecar {
      * Publish Error in Configured Metric/LogGroup/LogStream/LogConcentrator
      */
     sendResponse(response: APIGatewayProxyResult, callback: Callback): void {
-        
+
         this.publishOperationTrail()
         callback(null, response);
-        
+
     }
 
     /**
@@ -320,40 +419,7 @@ export class Sidecar {
      * @description
      * Match permissions of user with resource permissions
      */
-    private matchPermissions(): boolean {return true;}
-
-    /**
-     * @param error SidecarError
-     * @return InfrastructureMetric
-     * @description
-     * Create a Metric Object based on Error
-     */
-    private createMetricFromError(error: SidecarError): InfrastructureMetric {
-
-        const metric: InfrastructureMetric = {
-            metricArn: '',
-            value: { value: 1 }
-        }
-
-        return metric;
-
-    }
-
-    private gerErrorCode(error: SidecarError): number {
-
-        switch (error) {
-            case SidecarError.UserNotAuthorized:return 403;
-            case SidecarError.InvalidObjectBody: return 400;
-            case SidecarError.InvalidOperation: return 400;
-            case SidecarError.ResourceNotAvailable: return 500;
-            case SidecarError.DependencyError: return 500;
-            case SidecarError.ConfigurationNotAvailable: return 500;
-            case SidecarError.InvalidConfiguration: return 500;
-            default: return 500;
-        }
-
-    }
-
+    private matchPermissions(): boolean { return true; }
 
     /**
      * @param message Error message
@@ -363,6 +429,14 @@ export class Sidecar {
      * error metric could not be instantiated.
      */
     private publishOperationTrail(): void {
+
+        if (!isLambda) {
+            if (!existsSync('./testreports')) {mkdirSync('./testreports')}
+            const filename = new Date().getTime() + '.json'
+            writeFileSync('./testreports/lambdaexecution' + filename, JSON.stringify(this.trail, null, 2), 'utf-8');
+
+        }
+
         //console.log(this.trail);
     }
 
@@ -374,8 +448,17 @@ export class Sidecar {
      * error metric could not be instantiated.
      */
     private publishErrorInContingence(message: string): void {
-        console.error('Publishing Error in Contigence!')
-        console.error(message);
+
+        if (!isLambda) {
+
+            const filename = new Date().getTime() + '.json'
+            writeFileSync('./console_' + filename, JSON.stringify(this.trail, null, 2), 'utf-8');
+
+        } else {
+            console.error('Publishing Error in Contigence!')
+            console.error(message);
+        }
+
     }
 
 }
@@ -411,7 +494,7 @@ class OperationBuilder {
     payload?: string;
     error?: Object;
 
-    constructor() {this.timestamp = new Date();}
+    constructor() { this.timestamp = new Date(); }
 
     withAction(name: string): OperationBuilder { this.action = name; return this }
     withParameters(object: Object): OperationBuilder { this.parameters = object; return this }
@@ -423,7 +506,8 @@ class OperationBuilder {
         const now = new Date();
         this.durationInMillis = (now.getTime() - this.timestamp.getTime());
 
-        if (this.error === null || this.error === undefined) {this.success = true;
+        if (this.error === null || this.error === undefined) {
+            this.success = true;
         } else { this.success = false; }
 
         return new SidecarOperation(this);
